+++ cookbooks/vm/recipes/docker.rb
@@ -2,22 +2,26 @@
 # add the docker apt repository
 include_recipe 'chef-apt-docker'
 
-# install the docker-engine
-docker_installation_package 'default' do
-  version '1.11.0'
-  action :create
-  package_options "--force-yes -o Dpkg::Options::='--force-confold' -o Dpkg::Options::='--force-all'"
-end
-
-# start the docker deamon as a service
-docker_service_manager 'default' do
-  action :start
-end
-
 # create the docker group and add the devbox_user to it
 group 'docker' do
   action :create
   members devbox_user
   append true
-  notifies :restart, 'docker_service_manager[default]', :immediately
+end
+
+# FIXME: running docker-in-docker-in-docker still fails on circleci, so we skip
+# starting the deamon on circleci for now
+if docker?
+  # only install docker, don't try to start the deamon
+  docker_installation_package 'default' do
+    version '1.11.0'
+    action :create
+  end
+else
+  # install docker and start the docker deamon
+  docker_service 'default' do
+    install_method 'package'
+    version '1.11.0'
+    action [:create, :start]
+  end
 end