+++ cookbooks/vm/spec/integration/recipes/git_spec.rb
@@ -1,5 +1,4 @@
 require 'spec_helper'
-require 'chef/sugar/docker'
 
 describe 'vm::git' do